+++ apps/frontend/src/components/risikoscenario/edit/RisikoscenarioForm.tsx
@@ -3,9 +3,9 @@ import { Form, Formik } from 'formik'
 import { FunctionComponent, RefObject, useRef, useState } from 'react'
 import { mapRisikoscenarioToFormValue } from '../../../api/RisikoscenarioApi'
 import { IRisikoscenario } from '../../../constants'
-import { RisikoscenarioKonsekvensnivaa } from '../common/RisikoscenarioKonsekvensnivaa/RisikoscenarioKonsekvensnivaa'
 import RisikoscenarioSannsynlighet from '../common/RisikoscenarioSannsynlighet/RisikoscenarioSannsynlighet'
 import { RisikoscenarioBeskrivelseField } from './RisikoscenarioBeskrivelsField/RisikoscenarioBeskrivelseField'
+import { RisikoscenarioKonsekvensnivaaField } from './RisikoscenarioKonsekvensnivaaField/RisikoscenarioKonsekvensnivaaField'
 import { risikoscenarioCreateValidation } from './RisikoscenarioSchemaValidation'
 
 type TProps = {
@@ -41,7 +41,7 @@ export const RisikoscenarioForm: FunctionComponent<TProps> = ({
 
               <RisikoscenarioSannsynlighet />
 
-              <RisikoscenarioKonsekvensnivaa />
+              <RisikoscenarioKonsekvensnivaaField />
             </div>
 
             {Object.values(errors).some(Boolean) && (
