+++ apps/frontend/src/components/risikoscenario/edit/RisikoscenarioKonsekvensnivaaField/RisikoscenarioKonsekvensnivaaBegrunnelse.tsx
